Move Search.svelte into forms because it's a form. Implement viewer mode switching.
