+++ src/routes/app/+page.svelte
@@ -11,7 +11,7 @@
   import MainLayout from "$lib/components/layouts/MainLayout.svelte";
   import ContentLayout from "$lib/components/layouts/ContentLayout.svelte";
   import PageToolbar from "$lib/components/common/PageToolbar.svelte";
-  import Search from "$lib/components/inputs/Search.svelte";
+  import Search from "@/lib/components/forms/Search.svelte";
   import Empty from "$lib/components/common/Empty.svelte";
   import Button from "$lib/components/common/Button.svelte";
   import SignedIn from "$lib/components/common/SignedIn.svelte";
